requirements.txt
secretsharing
bitmerchant
blessings
versioneer
docopt
qrcode-terminal
